+++ migrations/frontend/1702500918_add_expiry_to_access_tokens/down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE access_tokens DROP COLUMN IF EXISTS expires_at;